apply 90f1aa253ca: [PATCH 08/13] Fix build on ARM architecture

+++ kexi/kexiutils/kmessagewidget.cpp
@@ -227,7 +227,7 @@ void KMessageWidgetPrivate::updateSnapShot()
 
 void KMessageWidgetPrivate::slotTimeLineChanged(qreal value)
 {
-    q->setFixedHeight(qMin(value * 2, 1.) * content->height());
+    q->setFixedHeight(qMin(value * 2, qreal(1.0)) * content->height());
     q->update();
 }
